requirements.txt
3to2==1.1.1
absl-py==0.11.0
allennlp==2.0.0
allennlp-models==1.4.0
appdirs==1.4.3
appnope==0.1.0
astroid==2.3.3
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
benepar==0.2.0a0
bleach==3.1.4
blis==0.4.1
bokeh==2.0.1
boto3==1.17.20
botocore==1.20.20
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.1
captum==0.4.0
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
check50==3.1.1
click==7.1.2
conllu==4.3
cryptography==2.9.2
cycler==0.10.0
cymem==2.0.3
datasets==1.2.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
distlib==0.3.0
docopt==0.6.2
en-core-web-md==2.2.5
en-core-web-sm==2.2.5
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
ftfy==5.9
future==0.18.2
gast==0.3.3
genbmm==0.1
gensim==3.8.3
google-auth==1.27.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
huggingface-hub==0.0.8
idna==2.10
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==3.4.0
iniconfig==1.1.1
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
jellyfish==0.8.2
Jinja2==2.11.1
jmespath==0.10.0
joblib==1.0.1
jsonnet==0.17.0
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lib50==2.0.8
lmdb==1.1.1
lxml==4.5.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.7.0
multiprocess==0.70.11.1
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.5
nltk==3.4.5
notebook==6.0.3
numpy==1.20.1
oauthlib==3.1.0
opt-einsum==3.3.0
overrides==3.1.0
packaging==20.9
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pipreqs==0.4.10
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.15.0rc1
ptyprocess==0.6.0
py==1.10.0
py-rouge==1.1
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.2.2
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytorch-transformers==1.1.0
pytz==2019.3
PyYAML==5.4.1
pyzmq==19.0.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
s3transfer==0.3.4
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.6.1
seaborn==0.10.0
Send2Trash==1.5.0
sentencepiece==0.1.95
six==1.15.0
sklearn-crfsuite==0.3.6
smart-open==3.0.0
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
torch==1.7.1
torch-struct==0.4
torch-tb-profiler==0.1.0
torchaudio==0.7.2
torchvision==0.8.2
tornado==6.0.4
tqdm==4.57.0
traitlets==4.3.3
transformers==4.6.0.dev0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.0.18
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
word2number==1.1
wrapt==1.12.1
xxhash==2.0.0
yarg==0.1.9
yellowbrick==1.2
zipp==3.4.0
